+++ guillotina/asgi.py
@@ -1,9 +1,11 @@
 from guillotina import glogging
 from guillotina import task_vars
+from guillotina.exc_resp import HTTPConflict
 from guillotina.exceptions import ConflictError
 from guillotina.exceptions import TIDConflictError
+from guillotina.middlewares import ErrorsMiddleware
 from guillotina.request import Request
-from guillotina.response import ASGISimpleResponse
+from guillotina.utils import resolve_dotted_name
 
 import asyncio
 import enum
@@ -20,13 +22,13 @@ class AppState(enum.IntEnum):
 
 
 class AsgiApp:
-    def __init__(self, config_file, settings, loop):
-        self.app = None
+    def __init__(self, config_file, settings, loop, router):
         self.config_file = config_file
         self.settings = settings
         self.loop = loop
+        self.router = router
+        self.app = None
         self.on_cleanup = []
-        self.route = None
         self.state = AppState.STARTING
 
     def __call__(self, scope, receive=None, send=None):
@@ -69,6 +71,7 @@ class AsgiApp:
             self.app = await startup_app(
                 config_file=self.config_file, settings=self.settings, loop=self.loop, server_app=self
             )
+            self.next_app = self.build_middleware_stack(self.app.settings)
             self.server_settings = self.app.settings.get("server_settings", {})
             self.state = AppState.INITIALIZED
             return self.app
@@ -83,6 +86,17 @@ class AsgiApp:
             await clean(self)
         self.state = AppState.SHUTDOWN
 
+    def build_middleware_stack(self, settings):
+        user_middlewares = [ErrorsMiddleware] + [
+            resolve_dotted_name(m) for m in settings.get("middlewares", [])
+        ]
+
+        # Guillotina is the last middleware in the chain
+        last_middleware = Guillotina(self.app, self.router)
+        for middleware in reversed(user_middlewares):
+            last_middleware = middleware(last_middleware)
+        return last_middleware
+
     async def handler(self, scope, receive, send):
         # Ensure the ASGI server has initialized the server before sending a request
         # Some ASGI servers (i.e. daphne) doesn't implement the lifespan protocol.
@@ -92,21 +106,39 @@ class AsgiApp:
         if scope["type"] == "websocket":
             scope["method"] = "GET"
 
-        request_settings = {k: v for k, v in self.server_settings.items() if k in ("client_max_size",)}
-        request = Request.factory(scope, send, receive, **request_settings)
-        task_vars.request.set(request)
-        resp = await self.request_handler(request)
+        resp = await self.next_app(scope, receive, send)
 
         if not resp.prepared:
+            request = task_vars.request.get()
             await resp.prepare(request)
+
+        return
+
+
+class Guillotina:
+    def __init__(self, asgi_app, router):
+        self.asgi_app = asgi_app
+        self.router = router
+
+    async def __call__(self, scope, receive, send):
+        request_settings = {
+            k: v for k, v in self.asgi_app.server_settings.items() if k in ("client_max_size",)
+        }
+        request = Request.factory(scope, send, receive, **request_settings)
+        task_vars.request.set(request)
+
+        resp = await self.request_handler(request)
+
+        return resp
 
     async def request_handler(self, request, retries=0):
         try:
-            route = await self.app.router.resolve(request)
-            return await route.handler(request)
+            route = await self.router.resolve(request)
+            resp = await route.handler(request)
+            return resp
 
         except (ConflictError, TIDConflictError) as e:
-            if self.app.settings.get("conflict_retry_attempts", 3) > retries:
+            if self.asgi_app.settings.get("conflict_retry_attempts", 3) > retries:
                 label = "DB Conflict detected"
                 if isinstance(e, TIDConflictError):
                     label = "TID Conflict Error detected"
@@ -133,4 +165,4 @@ class AsgiApp:
                         getattr(getattr(request, "_txn", None), "_tid", "not issued")
                     )
                 )
-                return ASGISimpleResponse(body=b"", status=409)
+                raise HTTPConflict()